+++ source/lib/output/generatePerfetto.hpp
@@ -44,9 +44,10 @@ write_perfetto(
     const generator<rocprofiler_buffer_tracing_hsa_api_record_t>&              hsa_api_gen,
     const generator<tool_buffer_tracing_kernel_dispatch_with_stream_record_t>& kernel_dispatch_gen,
     const generator<tool_buffer_tracing_memory_copy_with_stream_record_t>&     memory_copy_gen,
-    const generator<rocprofiler_buffer_tracing_marker_api_record_t>&           marker_api_gen,
-    const generator<rocprofiler_buffer_tracing_scratch_memory_record_t>&       scratch_memory_gen,
-    const generator<rocprofiler_buffer_tracing_rccl_api_record_t>&             rccl_api_gen,
+    const generator<tool_counter_record_t>&                                 counter_collection_gen,
+    const generator<rocprofiler_buffer_tracing_marker_api_record_t>&        marker_api_gen,
+    const generator<rocprofiler_buffer_tracing_scratch_memory_record_t>&    scratch_memory_gen,
+    const generator<rocprofiler_buffer_tracing_rccl_api_record_t>&          rccl_api_gen,
     const generator<rocprofiler_buffer_tracing_memory_allocation_record_t>& memory_allocation_gen,
     const generator<rocprofiler_buffer_tracing_rocdecode_api_record_t>&     rocdecode_api_gen,
     const generator<rocprofiler_buffer_tracing_rocjpeg_api_record_t>&       rocjpeg_api_gen);