+++ tests/rocprofv3/tracing-plus-counter-collection/validate.py
@@ -48,7 +48,9 @@ def test_perfetto_data(pftrace_data, json_data):
     import rocprofiler_sdk.tests.rocprofv3 as rocprofv3
 
     rocprofv3.test_perfetto_data(
-        pftrace_data, json_data, ("hip", "hsa", "marker", "kernel", "memory_copy")
+        pftrace_data,
+        json_data,
+        ("hip", "hsa", "marker", "kernel", "memory_copy", "counter_collection"),
     )