+++ CHANGELOG.md
@@ -171,6 +171,7 @@ Full documentation for ROCprofiler-SDK is available at [rocm.docs.amd.com/projec
 - Added MI350X/MI355X support
 - Added rocprofiler_create_counter to allow for adding custom derived counters at runtime.
 - Added support for iteration based counter multiplexing to rocprofv3 (see documentation)
+- Added perfetto support for counter collection.
 
 ### Changed